Disable ZIP compression, and comment out CRC32 checks, to improve loading times.
It turns out ZIP decompression is pretty slow, and the optional CRC32 check on data integrity is even slower (at least as implemented by miniz).  Long-term plan is to switch to LZ4 or some other compression method, see issue #3, but for now simply disabling compression and commenting out the CRC32 is easy and gets things loading quite a bit faster.